backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
python-dotenv>=0.21.0
httpx>=0.23.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
docx2txt>=0.8
pytesseract>=0.3.8
Pillow>=9.0.0
faiss-cpu>=1.7.0
numpy>=1.23.0
langchain>=0.0.300
langchain-community>=0.0.10
pydantic[email]>=2.0.0
python-dateutil>=2.8.0
typing-extensions>=4.5.0
